deleted MyHashMap(old version), renamed MyHashMap2 to MyHashMap
